sithafal/Task 1/requirements.txt
PyPDF2
langchain
sentence-transformers
faiss-cpu
transformers
pandas